_requirements.txt
aioredis==2.0.1
Flask==2.1.1
msal
PyJWT==1.7.1
pyodbc==4.0.32
redis>=4.1.4
requests>=2.25.1
botbuilder-core==4.14.2
botframework-connector==4.14.2
botframework-streaming==4.14.2
azure-storage-blob

aiohttp==3.7.4.post0
aiosmtplib==1.1.6
async-timeout>=3.0.1
botbuilder-schema==4.14.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==3.4.8
Deprecated==1.2.13
idna==2.10
importlib-metadata==4.11.3
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.1
jsonpickle==1.4.2
MarkupSafe==2.1.1
msrest==0.6.19
oauthlib==3.2.0
packaging==21.3
pycparser==2.21
pyparsing==3.0.8
requests-oauthlib==1.3.1
six==1.16.0
typing-extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.1.1
wrapt==1.14.0
zipp==3.8.0
